requirements.txt
requests
psutil
pyyaml